+++ src/enqueue_places.js
@@ -37,9 +37,13 @@ const enqueuePlacesFromResponse = (options) => {
             const pageNumber = parseInt(queryParams.ech, 10);
             // Parse place ids from response body
             const responseBody = await response.buffer();
-            const places = parseSearchPlacesResponseBody(responseBody);
+            const placesPaginationData = parseSearchPlacesResponseBody(responseBody);
             let index = -1;
-            for (const place of places) {
+            let enqueued = 0;
+            // At this point, page URL should be resolved
+            const searchPageUrl = page.url();
+
+            for (const placePaginationData of placesPaginationData) {
                 index++;
                 const rank = ((pageNumber - 1) * 20) + (index + 1);
                 if (!maxPlacesPerCrawl || rank <= maxPlacesPerCrawl) {
@@ -48,23 +52,33 @@ const enqueuePlacesFromResponse = (options) => {
                             url: `https://www.google.com/maps/search/?api=1&query=${searchString}&query_place_id=${place.placeId}`,
                         });
                     } else {
-                        const coordinates = allPlaces[place.placeId];
-                        if (!cachePlaces || !geo || !coordinates || checkInPolygon(geo, coordinates)) {
-                            // At this point, page URL should be resolved
-                            const searchPageUrl = page.url();
+                        // TODO: Refactor this once we get rid of the caching
+                        const coordinates = placePaginationData.coords || allPlaces[placePaginationData.placeId];
+                        const placeUrl = `https://www.google.com/maps/search/?api=1&query=${searchString}&query_place_id=${placePaginationData.placeId}`;
+                        if (!geo || !coordinates || checkInPolygon(geo, coordinates)) {
                             await requestQueue.addRequest({
-                                url: `https://www.google.com/maps/search/?api=1&query=${searchString}&query_place_id=${place.placeId}`,
-                                uniqueKey: place.placeId,
-                                userData: { label: 'detail', searchString, rank, searchPageUrl },
+                                url: placeUrl,
+                                uniqueKey: placePaginationData.placeId,
+                                userData: {
+                                    label: 'detail',
+                                    searchString,
+                                    rank,
+                                    searchPageUrl,
+                                    coords: placePaginationData.coords,
+                                    addressParsed: placePaginationData.addressParsed,
+                                },
                             },
                             { forefront: true });
+                            enqueued++;
                         } else {
-                            log.info(`[SEARCH]: Skip place: ${place.placeId}`);
                             stats.outOfPolygonCached();
+                            stats.outOfPolygon();
+                            stats.addOutOfPolygonPlace({ url: placeUrl, searchPageUrl, coordinates });
                         }
                     }
                 }
             }
+            log.info(`[SEARCH]: Enqueued ${enqueued}/${placesPaginationData.length} places (correct location/total) --- ${page.url()}`)
         }
     };
 };